[BUG] Validate dtype when Int64Index, UInt64Index, or Float64Index are cons… (#29545)

+++ pandas/core/indexes/base.py
@@ -381,7 +381,7 @@ class Index(IndexOpsMixin, PandasObject):
                             pass
 
                         # Return an actual float index.
-                        return Float64Index(data, copy=copy, dtype=dtype, name=name)
+                        return Float64Index(data, copy=copy, name=name)
 
                     elif inferred == "string":
                         pass
